feat: split up program in different h files, implement include guards so that STM32CubeIDE project can use this code, make weights generation script write to a bin file instead of a .h file
