+++ include/LIF_neuron.h
@@ -1,0 +1,7 @@
+#pragma once
+////////////////////// LIFNeuron struct //////////////////////////////
+typedef struct {
+    double mp; // membrane potential
+    double ts; // time-stamp of last spike
+} LIFNeuron;
+//////////////////////////////////////////////////////////////////////
