+++ include/weights.h
@@ -1,0 +1,16 @@
+#pragma once
+
+weights[0][0][0] = 0.5837970086175467;
+weights[0][1][0] = 0.6420932531211204;
+weights[1][0][1] = 0.7341234949183775;
+weights[1][1][0] = 0.4372310817050329;
+weights[1][2][0] = 0.5150265320077378;
+weights[0][0][0] = 0.14652514764942093;
+weights[0][1][1] = 0.7158070002609322;
+weights[0][1][2] = 0.2594500864888907;
+weights[1][1][0] = 0.7237455691313709;
+weights[1][1][1] = 0.16811820286144485;
+weights[0][0][1] = 0.8632614685523236;
+weights[1][0][0] = 0.7980893249210784;
+weights[1][0][1] = 0.22059113796665264;
+weights[1][1][0] = 0.4757397605258039;
